move to using Verify
