+++ src/Directory.build.props
@@ -41,7 +41,7 @@
     <PackageReference Include="FluentAssertions" Version="6.1.0" />
     <PackageReference Include="Microsoft.Reactive.Testing" Version="5.0.0" />
     <PackageReference Include="PublicApiGenerator" Version="10.2.0" />
-    <PackageReference Include="DiffEngine" Version="7.3.0" />
+    <PackageReference Include="Verify.Xunit" Version="12.2.0" />
   </ItemGroup>
 
   <ItemGroup Condition="$(IsTestProject)">
