+++ src/Splat.Tests/API/ApiApprovalTests.SplatProject.DotNet6_0.verified.txt
