+++ src/Splat.Tests/API/ApiApprovalTests.cs
@@ -5,8 +5,15 @@
 
 #if !WINDOWS_UWP && !ANDROID
 
+using System;
 using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using PublicApiGenerator;
+using VerifyXunit;
 using Xunit;
+#pragma warning disable SA1615 // Element return value should be documented
 
 namespace Splat.Tests
 {
@@ -14,15 +21,38 @@ namespace Splat.Tests
     /// Tests to make sure that the API matches the approved ones.
     /// </summary>
     [ExcludeFromCodeCoverage]
+    [UsesVerify]
     public class ApiApprovalTests
     {
+        private static readonly Regex _removeCoverletSectionRegex = new(@"^namespace Coverlet\.Core\.Instrumentation\.Tracker.*?^}", RegexOptions.Singleline | RegexOptions.Multiline | RegexOptions.Compiled);
+
         /// <summary>
         /// Tests to make sure the splat project is approved.
         /// </summary>
         [Fact]
-        public void SplatProject()
+        public Task SplatProject()
         {
-            typeof(AssemblyFinder).Assembly.CheckApproval();
+            var assembly = typeof(AssemblyFinder).Assembly;
+            var generatorOptions = new ApiGeneratorOptions { WhitelistedNamespacePrefixes = new[] { "Splat" } };
+            var receivedPublicApi = Filter(assembly.GeneratePublicApi(generatorOptions));
+            return Verifier.Verify(receivedPublicApi).UniqueForRuntimeAndVersion();
+        }
+
+        private static string Filter(string text)
+        {
+            text = _removeCoverletSectionRegex.Replace(text, string.Empty);
+            return string.Join(Environment.NewLine, text.Split(
+                    new[]
+                    {
+                        Environment.NewLine
+                    },
+                    StringSplitOptions.RemoveEmptyEntries)
+                .Where(l =>
+                    !l.StartsWith("[assembly: AssemblyVersion(", StringComparison.InvariantCulture) &&
+                    !l.StartsWith("[assembly: AssemblyFileVersion(", StringComparison.InvariantCulture) &&
+                    !l.StartsWith("[assembly: AssemblyInformationalVersion(", StringComparison.InvariantCulture) &&
+                    !l.StartsWith("[assembly: System.Reflection.AssemblyMetadata(", StringComparison.InvariantCulture) &&
+                    !string.IsNullOrWhiteSpace(l)));
         }
     }
 }
